api: Roll back some changes; adjust year handling
The container option, reparse on init, and yearsComplete callback were
pretty unnecessary, so removed those.

The render function now takes a year as an argument, and optionally adds
an "all years" item to the template data.

Also added support for a year selectbox in addition to link triggers.

Plus renamed "item" functions to "result".